requirements.txt
numpy==1.26.4
scipy==1.13.0
pyserial==3.5
matplotlib==3.9.1
tk==0.1.0
requests==2.32.3
pygame==2.6.0
scikit-learn==1.5.1
fs==2.4.16
brainflow==5.13.3
mne==1.8.0
pyserial==3.5